Add option to handle EOLs differently.
By default handle EOLs in a specific way that makes it easier to verify the files from an archive because all files have the same EOL as a tgz archive now (almost all LF). The remaining files unique to a ZIP also keep the same EOL (CRLF).

The option to handle EOLs automatically is still available.